Moved 2 files Expanded the commentary somewhat. Signed-off-by: Joey Smith <jsmith@webinertia.net>
