+++ module/App/src/Message.php
@@ -2,7 +2,9 @@
 
 declare(strict_types=1);
 
-namespace App\Model;
+namespace App;
+
+use App\Model\AbstractMessage;
 
 final class Message extends AbstractMessage
 {
